camp/week_5/telebot/requirements.txt
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
pycodestyle==2.10.0
pyTelegramBotAPI==4.11.0
python-decouple==3.8
requests==2.30.0
tomli==2.0.1
urllib3==2.0.2
